Split certificate module to multiple files
