+++ tedge/src/cli/certificate/mod.rs
@@ -1,0 +1,8 @@
+pub use self::cli::TEdgeCertCli;
+
+mod cli;
+mod create;
+mod error;
+mod remove;
+mod show;
+mod upload;
