+++ tedge/src/main.rs
@@ -4,7 +4,6 @@
 use anyhow::Context;
 use structopt::StructOpt;
 
-mod certificate;
 mod cli;
 mod command;
 mod error;
